reactive_graph_std_connector::behaviour::relation::connector

Module function

Constants

FN_DEBUG_CONNECTOR

This connector logs the value before propagation (log level debug)

FN_DEFAULT_CONNECTOR

This is the default connector type, which simply does nothing than propagate the value

FN_PARSE_FLOAT_CONNECTOR

This connector parses a string value and propagates a float value

FN_PARSE_INT_CONNECTOR

This connector parses a string value and propagates a int value

FN_TO_STRING_CONNECTOR

This connector converts the value of any type to string before propagation

FN_TRACE_CONNECTOR

This connector logs the value before propagation (log level trace)

Statics

CONNECTOR_BEHAVIOURS

Type Aliases

ConnectorFunction